requirements.txt
streamlit
tensorflow
pillow
scikit-image
numpy
torch
torchvision
git+https://github.com/richzhang/PerceptualSimilarity
